Add UUID generation for windows, fix ms timestamp overflow
an unsigned long isn't large enough to hold a ms timestamp in some
situations, I have changed this to be an unsigned long long. This too
may not be large enough in some situations (system clock epoch is not
currently set in stone as jan 1 1970 - it could be much earlier).
However for now people seem to have settled on jan 1 1970, so I think
this fix is ok for now.